+++ packages/database/lib/src/database.dart
@@ -17,6 +17,7 @@ part 'daos/transactions_dao.dart';
 part 'daos/categories_dao.dart';
 part 'daos/budgets_dao.dart';
 part 'daos/import_sources_dao.dart';
+part 'daos/tags_dao.dart';
 
 /// Local finance database with all tables.
 @DriftDatabase(
@@ -45,6 +46,7 @@ class LocalFinanceDatabase extends _$LocalFinanceDatabase {
   late final CategoriesDao categoriesDao = CategoriesDao(this);
   late final BudgetsDao budgetsDao = BudgetsDao(this);
   late final ImportSourcesDao importSourcesDao = ImportSourcesDao(this);
+  late final TagsDao tagsDao = TagsDao(this);
 
   @override
   int get schemaVersion => 5;